Feature add trips pages

+++ app/Http/Controllers/Api/CampaignsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Api;
+namespace App\Http\Controllers\v1;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests;
@@ -36,6 +36,10 @@ class CampaignsController extends Controller
      */
     public function index(Request $request)
     {
+        if ( ! $this->auth->user()->isAdmin()) {
+            abort(403);
+        }
+
         $campaigns = $this->campaign
                           ->filter($request->all())
                           ->paginate($request->get('per_page', 25));
@@ -51,6 +55,10 @@ class CampaignsController extends Controller
      */
     public function show($id)
     {
+        if ( ! $this->auth->user()->isAdmin()) {
+            abort(403);
+        }
+
         $campaign = $this->campaign->findOrFail($id);
 
         return $this->response->item($campaign, new CampaignTransformer);
@@ -93,6 +101,10 @@ class CampaignsController extends Controller
      */
     public function destroy($id)
     {
+        if ( ! $this->auth->user()->isAdmin()) {
+            abort(403);
+        }
+
         $campaign = $this->campaign->findOrFail($id);
 
         $campaign->delete();
